UI-50 rename files
