--- src/Date/index.ts
@@ -1,1 +1,0 @@
-export * from './Date'
