+++ src/index.ts
@@ -24,7 +24,7 @@ export * from './RadioGroup'
 export * from './SearchInput'
 export * from './Footer'
 export * from './DateInput'
-export * from './Date'
+export * from './DateTextInput'
 export * from './Spinner'
 export * from './Toggle'
 export * from './style/colors'
